# tests/api/golden/openapi.json
{
  "components": {
    "schemas": {
      "Attachment": {
        "additionalProperties": false,
        "description": "Attachment as it appears on an inbound or outbound envelope.\n\nURLs are always adapter-hosted on inbound (spec §7.3.1). For the outbound\n*input* shape (where the caller may supply a local file path instead of a\nURL), use :class:`SendAttachmentInput` on :class:`SendMessageRequest`.",
        "properties": {
          "id": {
            "pattern": "^att_[0-9A-HJKMNP-TV-Z]{26}$",
            "title": "Id",
            "type": "string"
          },
          "mime": {
            "minLength": 1,
            "title": "Mime",
            "type": "string"
          },
          "size_bytes": {
            "minimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          },
          "url": {
            "description": "Adapter-hosted HTTP(S) URL.",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "mime",
          "size_bytes",
          "url"
        ],
        "title": "Attachment",
        "type": "object"
      },
      "ChannelType": {
        "description": "Channel topology. v1 returns only ``dm`` for iMessage; both for Discord.",
        "enum": [
          "dm",
          "group"
        ],
        "title": "ChannelType",
        "type": "string"
      },
      "Direction": {
        "description": "Which way a message moved through AMC.",
        "enum": [
          "inbound",
          "outbound"
        ],
        "title": "Direction",
        "type": "string"
      },
      "Envelope": {
        "additionalProperties": false,
        "description": "Normalized cross-platform message envelope (spec §7.3.1).\n\nEvery inbound and outbound message conforms to this shape. New platforms\nare added by writing one connector that produces an :class:`Envelope`;\nnothing else in the system needs to change.",
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/Attachment"
            },
            "title": "Attachments",
            "type": "array"
          },
          "channel_id": {
            "minLength": 1,
            "title": "Channel Id",
            "type": "string"
          },
          "channel_type": {
            "$ref": "#/components/schemas/ChannelType"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          },
          "id": {
            "pattern": "^msg_[0-9A-HJKMNP-TV-Z]{26}$",
            "title": "Id",
            "type": "string"
          },
          "raw": {
            "additionalProperties": true,
            "description": "Platform-native payload for debugging / forward-compat.",
            "title": "Raw",
            "type": "object"
          },
          "reply_to": {
            "anyOf": [
              {
                "pattern": "^msg_[0-9A-HJKMNP-TV-Z]{26}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Another msg_... id within the same channel, if threaded.",
            "title": "Reply To"
          },
          "sender": {
            "$ref": "#/components/schemas/Sender"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          },
          "text": {
            "description": "UTF-8; empty allowed for attachment-only.",
            "title": "Text",
            "type": "string"
          },
          "timestamp": {
            "description": "Platform-claimed message time (RFC 3339 UTC).",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "channel_id",
          "channel_type",
          "direction",
          "id",
          "sender",
          "source",
          "text",
          "timestamp"
        ],
        "title": "Envelope",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Sender": {
        "additionalProperties": false,
        "description": "Sender identity carried inside the envelope (spec §7.3.1).",
        "properties": {
          "display_name": {
            "description": "Allowlist override else platform value.",
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "description": "Platform-native sender id.",
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "person_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present iff this sender is allowlisted with a person_id.",
            "title": "Person Id"
          }
        },
        "required": [
          "display_name",
          "id"
        ],
        "title": "Sender",
        "type": "object"
      },
      "Source": {
        "description": "Originating platform for a message.",
        "enum": [
          "discord",
          "imessage"
        ],
        "title": "Source",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Agent Messaging Channel adapter — a single-Mac REST surface that normalizes iMessage and Discord behind one bearer-protected API. All endpoints, including this OpenAPI document and the Swagger UI, require `Authorization: Bearer <token>` per spec §6.2 and §7.4.",
    "title": "AMC Adapter",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/messages/context": {
      "get": {
        "description": "``GET /messages/context`` — return chronologically ordered context.\n\nBearer + ``X-Agent-ID`` are enforced by the router-level ``dependencies``.\n``before`` and ``after`` are clamped at ``[0, 50]`` by Pydantic; values\noutside the range surface as the standard ``VALIDATION_FAILED`` envelope\nvia the global handler in :mod:`amc.core.errors`.",
        "operationId": "get_messages_context_messages_context_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Agent-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Agent-Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 50,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "around_message_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Around Message Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 50,
              "minimum": 0,
              "title": "Before",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "channel_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Channel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Messages Context"
      }
    },
...
    "/messages/{message_id}": {
      "get": {
        "description": "Return the single message identified by ``message_id``.\n\nVisibility rules:\n\n* ``allowlist_status='allowed'`` → visible (typical inbound).\n* ``allowlist_status='outbound'`` → visible (agent-sent).\n* ``allowlist_status='unknown'`` → hidden (return 404).\n* Row missing → 404.\n\nThe two \"not visible\" cases collapse to the same 404 envelope so the\ncaller cannot distinguish quarantined-but-present from truly-absent.",
        "operationId": "get_message_messages__message_id__get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Agent-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Agent-Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "description": "ULID-prefixed message id (``msg_<26 Crockford-base32 chars>``).",
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "description": "ULID-prefixed message id (``msg_<26 Crockford-base32 chars>``).",
              "pattern": "^msg_[0-9A-HJKMNP-TV-Z]{26}$",
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Message"
      }
    }
  }
}